requirements.txt
pandas
python-dateutil
PyMySQL
SQLAlchemy
tushare
